+++ Modules/Ruleset.ProgramInfo/Public/Test-ExecutableFile.ps1
@@ -34,7 +34,7 @@ Check if executable file exists and is trusted.
 Test-ExecutableFile verifies the path to executable file is valid and that executable itself exists.
 File extension is then verified to confirm it is whitelisted, ex. such as an *.exe
 The executable is then verified to ensure it's digitaly signed and that signature is valid.
-If digital signature is missing or not valid, the file is optionally scanned on virus total to
+If digital signature is missing or not valid, the file is optionally scanned on VirusTotal to
 confirm it's not malware.
 If the file can't be found or verified, an error is genrated possibly with informational message,
 to explain if there is any problem with the path or file name syntax, otherwise information is
@@ -58,10 +58,10 @@ Specify path to sigcheck executable program.
 Do not specify sigcheck file, only path to where sigcheck is located.
 By default working directory and PATH is searched for sigcheck64.exe.
 On 32 bit operating system sigcheck.exe is searched instead.
-If location to sigcheck executable is not found then no virus total scan and report is done.
+If location to sigcheck executable is not found then no VirusTotal scan and report is done.
 
 .PARAMETER Timeout
-Specify maximum wait time expressed in seconds for virus total to scan individual file.
+Specify maximum wait time expressed in seconds for VirusTotal to scan individual file.
 Value 0 means an immediate return, and a value of -1 specifies an infinite wait.
 The default wait time is 300 (5 minutes).
 
@@ -71,7 +71,7 @@ If specified, no information, warning or error message is shown, only true or fa
 .PARAMETER Force
 If specified, lack of digital signature or signature mismatch produces a warning
 instead of an error resulting in bypassed signature test.
-This parameter has no effect virus total check, if the file is reported as malware the return value is False.
+This parameter has no effect VirusTotal check, if the file is reported as malware the return value is False.
 
 .EXAMPLE
 PS> Test-ExecutableFile "C:\Windows\UnsignedFile.exe"